+++ examples/gradient-embassy/src/main.rs
@@ -49,6 +49,7 @@ use esp_hal::time::Rate;
 use esp_hal::timer::timg::TimerGroup;
 use esp_hub75::Color;
 use esp_hub75::Hub75;
+use esp_hub75::Hub75Config;
 use esp_hub75::Hub75Pins16;
 #[cfg(not(feature = "row"))]
 use esp_hub75::framebuffer::bitplane::plain::DmaFrameBuffer;
@@ -324,7 +325,7 @@ async fn main(spawner: Spawner) {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -335,7 +336,7 @@ async fn main(spawner: Spawner) {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -348,7 +349,7 @@ async fn main(spawner: Spawner) {
         pins,
         peripherals.DMA_I2S0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");