+++ examples/gradient-quarter/src/main.rs
@@ -56,6 +56,7 @@ use esp_hal::time::Instant;
 use esp_hal::time::Rate;
 use esp_hub75::Color;
 use esp_hub75::Hub75;
+use esp_hub75::Hub75Config;
 use esp_hub75::Hub75Pins16;
 use esp_hub75::framebuffer::FrameBufferOperations;
 #[cfg(not(feature = "row"))]
@@ -274,7 +275,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -285,7 +286,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -298,7 +299,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_I2S0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");