+++ examples/gradient-tiled/src/main.rs
@@ -53,6 +53,7 @@ use esp_hal::time::Instant;
 use esp_hal::time::Rate;
 use esp_hub75::Color;
 use esp_hub75::Hub75;
+use esp_hub75::Hub75Config;
 use esp_hub75::Hub75Pins16;
 use esp_hub75::framebuffer::FrameBufferOperations;
 #[cfg(not(feature = "row"))]
@@ -263,7 +264,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -274,7 +275,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -287,7 +288,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_I2S0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");