+++ examples/gradient/src/main.rs
@@ -44,6 +44,7 @@ use esp_hal::time::Instant;
 use esp_hal::time::Rate;
 use esp_hub75::Color;
 use esp_hub75::Hub75;
+use esp_hub75::Hub75Config;
 use esp_hub75::Hub75Pins16;
 #[cfg(not(feature = "row"))]
 use esp_hub75::framebuffer::bitplane::plain::DmaFrameBuffer;
@@ -220,7 +221,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -231,7 +232,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_CH0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");
@@ -244,7 +245,7 @@ fn main() -> ! {
         pins,
         peripherals.DMA_I2S0,
         tx_descriptors,
-        rate,
+        Hub75Config::new(rate),
         &*fb0,
     )
     .expect("failed to create Hub75");